Initial configuration for the views
