+++ gateway/api/views/__init__.py
@@ -1,0 +1,10 @@
+"""
+Django Rest framework ViewSets for api application:
+
+Version views inherit from the different views.
+"""
+
+from .catalog import CatalogViewSet
+from .files import FilesViewSet
+from .jobs import JobViewSet
+from .programs import ProgramViewSet
